+++ src/dra/menu.c
@@ -4093,38 +4093,3 @@ void func_800FD39C(s32 x, s32 y, s32 w, s32 h, s32 u, s32 v, s32 pal, s32 _,
     g_GpuUsage.sp++;
 }
 #endif
-
-s32 TimeAttackController(TimeAttackEvents eventId, TimeAttackActions action) {
-    s32 temp_v0;
-    s32 temp_v1;
-    s32 seconds;
-    s32 timer;
-
-    switch (action) {
-    case TIMEATTACK_GET_RECORD:
-        // get the time attack for a specific defeated boss. this is
-        // also responsible to check if the player should teleport into
-        // a boss room
-        return g_Settings.timeAttackRecords[eventId];
-
-    case TIMEATTACK_SET_RECORD:
-        // set new time attack record if the boss was not previously
-        // defeated
-        timer = g_Settings.timeAttackRecords[eventId];
-        if (timer = timer != 0) {
-            return g_Settings.timeAttackRecords[eventId];
-        }
-
-        seconds = g_Status.timerSeconds;
-        g_Settings.timeAttackRecords[eventId] = seconds;
-        temp_v1 = (g_Status.timerMinutes * 100) + seconds;
-        g_Settings.timeAttackRecords[eventId] = temp_v1;
-        temp_v0 = (g_Status.timerHours * 10000) + temp_v1;
-        g_Settings.timeAttackRecords[eventId] = temp_v0;
-        return temp_v0;
-
-    case TIMEATTACK_SET_VISITED:
-        // not exactly sure yet why this flag is needed
-        g_Settings.D_8003CB00 |= 1 << eventId;
-    }
-}
